requirements.txt
django==3.2.5
djangorestframework==3.13.1 
psycopg2==2.8.6
python-dotenv==0.19.2 